[JSONRPC] Seperate data fetching from RPC implementation
Closes: #8278
